@@ -1,24 +1,28 @@
 module github.com/Luzifer/share
 
-go 1.20
+go 1.22.1
 
 require (
-	github.com/Luzifer/rconfig/v2 v2.4.0
-	github.com/aws/aws-sdk-go v1.44.275
-	github.com/cheggaaa/pb v1.0.29
+	github.com/Luzifer/rconfig/v2 v2.5.0
+	github.com/aws/aws-sdk-go v1.51.1
+	github.com/cheggaaa/pb/v3 v3.1.5
 	github.com/gofrs/uuid v4.4.0+incompatible
-	github.com/gosimple/slug v1.13.1
+	github.com/gosimple/slug v1.14.0
 	github.com/pkg/errors v0.9.1
-	github.com/rivo/uniseg v0.4.4 // indirect
 	github.com/sirupsen/logrus v1.9.3
 )
 
 require (
+	github.com/VividCortex/ewma v1.2.0 // indirect
+	github.com/fatih/color v1.15.0 // indirect
 	github.com/gosimple/unidecode v1.0.1 // indirect
 	github.com/jmespath/go-jmespath v0.4.0 // indirect
-	github.com/mattn/go-runewidth v0.0.14 // indirect
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.19 // indirect
+	github.com/mattn/go-runewidth v0.0.15 // indirect
+	github.com/rivo/uniseg v0.2.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
-	golang.org/x/sys v0.8.0 // indirect
+	golang.org/x/sys v0.6.0 // indirect
 	gopkg.in/validator.v2 v2.0.1 // indirect
-	gopkg.in/yaml.v2 v2.4.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
